requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
astroid==2.15.7
backports.zoneinfo==0.2.1
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.7
Django==3.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg==3.1.11
psycopg2==2.9.7
pydantic==2.4.0
pydantic_core==2.10.0
PyJWT==1.7.1
pylint==2.17.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing-inspect==0.8.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.5
wrapt==1.15.0
